pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.116.1",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.27",
    "libmagic>=1.0",
    "openai>=1.95.1",
    "python-dotenv>=1.1.1",
    "python-magic>=0.4.27",
    "unstructured[md]>=0.18.5",
    "uvicorn>=0.35.0",
]
